pilot edits after 2nd round of feedback
