+++ droidlet/tools/crowdsourcing/vision_annotation_task/conf/pilot.yaml
@@ -11,12 +11,12 @@ mephisto:
     task_description: "Mark the blocks that make up a shape"
     task_reward: 2.5
     allowed_concurrent: 1
-    maximum_units_per_worker: 1
+    # maximum_units_per_worker: 1
     task_tags: "3D, annotate, shapes"
     task_name: "vision-pre-pilot-12-29-21-t1"
-  architect:
-    heroku_config_args:
-      AWS_ACCESS_KEY_ID: ${env:AWS_ACCESS_KEY_ID}
-      AWS_SECRET_ACCESS_KEY: ${env:AWS_SECRET_ACCESS_KEY}
-      AWS_DEFAULT_REGION: "us-west-1"
-    heroku_app_name: vision-pre-pilot-12-29-21-t1
+  # architect:
+  #   heroku_config_args:
+  #     AWS_ACCESS_KEY_ID: ${env:AWS_ACCESS_KEY_ID}
+  #     AWS_SECRET_ACCESS_KEY: ${env:AWS_SECRET_ACCESS_KEY}
+  #     AWS_DEFAULT_REGION: "us-west-1"
+  #   heroku_app_name: vision-pre-pilot-12-29-21-t1
