+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/extra_refs/pilot_utils.js
@@ -8,9 +8,10 @@
 $('link[rel=stylesheet][href~="https://cdn.jsdelivr.net/npm/bulma@0.8.2/css/bulma.min.css"]').remove();
 
 var clickedElements = new Array();
-var complete = [false, false, false]; 
-var correct = [false, false, false];
+var complete = [false, false, false, false]; 
+var correct = [false, false, false, false];
 var answers = [
+  [[-2,0,2], [-2,1,2], [-2,2,2]],
   [[-5,1,0],[-4,1,0],[-5,0,0],[-4,0,0],[-4,1,-1],[-5,1,-1],[-4,0,-1],[-5,0,-1]],
   null,
   [[-1,4,0],[0,4,0],[0,4,-1],[-1,4,-1]]
@@ -47,6 +48,7 @@ function recordClick(ele) {
   if (ele["q1_output"]) { checkAnswer(1, ele.q1_output) }
   else if (ele["q2_output"]) { checkAnswer(2, ele.q2_output) }
   else if (ele["q3_output"]) { checkAnswer(3, ele.q3_output) }
+  else if (ele["q4_output"]) { checkAnswer(4, ele.q4_output) }
 
   clickedElements.push({id: ele, timestamp: Date.now()});
   document.getElementById("clickedElements").value = JSON.stringify(clickedElements);
@@ -118,3 +120,23 @@ function showHideInstructions() {
   recordClick("toggle-instructions");
 }
 
+let font_size = 11;
+let heading_size = 13;
+function dynamicInstructionsSize() {
+  while (document.getElementById("heading").offsetHeight > (window.innerHeight * 0.9)) {
+    console.log("decrease font size");
+    font_size -= 1;
+    heading_size -= 1;
+    Array.from(document.getElementsByClassName("instructions-section")).forEach( ele => ele.style.fontSize = font_size + "pt" );
+    Array.from(document.getElementsByClassName("instruction-headings")).forEach( ele => ele.style.fontSize = heading_size + "pt" );
+  }
+  while (document.getElementById("heading").offsetHeight < (window.innerHeight * 0.8)) {
+    console.log("increase font size");
+    font_size += 1;
+    heading_size += 1;
+    Array.from(document.getElementsByClassName("instructions-section")).forEach( ele => ele.style.fontSize = font_size + "pt" );
+    Array.from(document.getElementsByClassName("instruction-headings")).forEach( ele => ele.style.fontSize = heading_size + "pt" );
+  }
+}
+dynamicInstructionsSize(); // Call once on page load
+window.addEventListener('resize', dynamicInstructionsSize);  // And on resize
