+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/extra_refs/voxel_painter_q1.mjs
@@ -1,6 +1,5 @@
 import * as THREE from 'https://cdn.skypack.dev/three';
 import { OrbitControls } from './OrbitControls.mjs';
-import { GLTFLoader } from './GLTFLoader.mjs';
 
 let camera1, controls1, scene1, renderer1, plane1;
 let camera2, controls2, scene2, renderer2, plane2;
@@ -14,47 +13,21 @@ let rollOverMesh2, rollOverMaterial2;
 let cubeMaterial_mark;
 const geo = new THREE.BoxGeometry( 50, 50, 50 );
 const cubeMaterial = new THREE.MeshLambertMaterial( { color: 0xfeb74c, map: new THREE.TextureLoader().load( 'square-outline-textured.png' ) } );
-const userMaterial = new THREE.MeshLambertMaterial( { color: 0xffff00 } );
-const agentMaterial = new THREE.MeshLambertMaterial( { color: 0x0000ff } );
 
 let objects1 = [];
 let objects2  = [];
 let marked_blocks = [];
 
-let userModel1, userModel2, agentModel1, agentModel2;
-
-let user_pos = [-6, -3];  // x,z
-let agent_pos = [6, 2];
-
 let actions_taken = []; // [original_block, new_block, action_type]
 var startedHIT = false;
 
 let starting_shapes = [
-    [5,0,0,0],
-    [5,1,0,1],
-    [4,0,0,2],
-    [4,1,0,3],
-    [5,0,-1,4],
-    [5,1,-1,5],
-    [4,0,-1,0],
-    [4,1,-1,1],
-    [0,0,0,0],
-    [0,1,0,1],
-    [-1,0,0,2],
-    [-1,1,0,3],
-    [0,0,-1,4],
-    [0,1,-1,5],
-    [-1,0,-1,0],
-    [-1,1,-1,1],
-    [-5,0,0,0],
-    [-5,1,0,1],
-    [-4,0,0,2],
-    [-4,1,0,3],
-    [-5,0,-1,4],
-    [-5,1,-1,5],
-    [-4,0,-1,0],
-    [-4,1,-1,1]
+    [[2,0,2,0], [2,1,2,0], [2,2,2,0]],
+    [[2,0,-2,1], [2,1,-2,0], [2,2,-2,0]],
+    [[-2,0,2,2], [-2,1,2,0], [-2,2,2,0]],
+    [[-2,0,-2,3], [-2,1,-2,0], [-2,2,-2,0]]
 ];
+let colors = [0xff0000, 0x00ff00, 0x0000ff, 0x000000]
 
 init1();
 init2();
@@ -85,41 +58,19 @@ function init1() {
     objects1.push( plane1 );
 
     // starting shapes
-    starting_shapes.forEach((shape) => {
-        const voxel = new THREE.Mesh( geo, cubeMaterial );
-        voxel.position.set((shape[0]*50)+25, (shape[1]*50)+25, (shape[2]*50)+25);
-        scene1.add( voxel );
-        objects1.push( voxel );
-    })
-
-    // add user
-    const loader = new GLTFLoader();
-    loader.load( './body.glb', function ( gltf ) {
-        userModel1 = gltf.scene;
-        userModel1.scale.multiplyScalar(75.0);
-        userModel1.position.set((user_pos[0]*50)+25, 0, (user_pos[1]*50)+25)
-        scene1.add( userModel1 );
-        userModel1.traverse( function ( object ) {
-            if ( object.isMesh ) {
-                object.castShadow = false;
-                object.material = userMaterial;
-            }
-        } );
-	} );
-
-    // add agent
-    loader.load( './body.glb', function ( gltf ) {
-        agentModel1 = gltf.scene;
-        agentModel1.scale.multiplyScalar(75.0);
-        agentModel1.position.set((agent_pos[0]*50)+25, 0, (agent_pos[1]*50)+25)
-        scene1.add( agentModel1 );
-        agentModel1.traverse( function ( object ) {
-            if ( object.isMesh ) {
-                object.castShadow = false;
-                object.material = agentMaterial;
-            }
-        } );
-	} );
+    for (let i=0; i<starting_shapes.length; i++) {
+        starting_shapes[i].forEach(block => {
+            const material = new THREE.MeshBasicMaterial( { color: colors[i], opacity: 1.0} );
+            const voxel = new THREE.Mesh( geo, material );
+            voxel.position.set((block[0]*50)+25, (block[1]*50)+25, (block[2]*50)+25);
+            scene1.add( voxel );
+            objects1.push( voxel );
+            const edges = new THREE.EdgesGeometry( geo );  // outline the blocks for visibility
+            const line = new THREE.LineSegments( edges, new THREE.LineBasicMaterial( { color: 0x000000 } ) );
+            line.position.set((block[0]*50)+25, (block[1]*50)+25, (block[2]*50)+25);
+            scene1.add( line );
+        })
+    }
 
     // lights
     const ambientLight = new THREE.AmbientLight( 0x606060 );
@@ -176,42 +127,20 @@ function init2() {
     scene2.add( plane2 );
     objects2.push( plane2 );
 
-    // starting shapes
-    starting_shapes.forEach((shape) => {
-        const voxel = new THREE.Mesh( geo, cubeMaterial );
-        voxel.position.set((shape[0]*50)+25, (shape[1]*50)+25, (shape[2]*50)+25);
-        scene2.add( voxel );
-        objects2.push( voxel );
-    });
-
-    // add user
-    const loader = new GLTFLoader();
-    loader.load( './body.glb', function ( gltf ) {
-        userModel2 = gltf.scene;
-        userModel2.scale.multiplyScalar(75.0);
-        userModel2.position.set((user_pos[0]*50)+25, 0, (user_pos[1]*50)+25)
-        scene2.add( userModel2 );
-        userModel2.traverse( function ( object ) {
-            if ( object.isMesh ) {
-                object.castShadow = false;
-                object.material = userMaterial;
-            }
-        } );
-    } );
-
-    // add agent
-    loader.load( './body.glb', function ( gltf ) {
-        agentModel2 = gltf.scene;
-        agentModel2.scale.multiplyScalar(75.0);
-        agentModel2.position.set((agent_pos[0]*50)+25, 0, (agent_pos[1]*50)+25)
-        scene2.add( agentModel2 );
-        agentModel2.traverse( function ( object ) {
-            if ( object.isMesh ) {
-                object.castShadow = false;
-                object.material = agentMaterial;
-            }
-        } );
-    } );
+    // starting shape
+    for (let i=0; i<starting_shapes.length; i++) {
+        starting_shapes[i].forEach(block => {
+            const material = new THREE.MeshBasicMaterial( { color: colors[i], opacity: 1.0} );
+            const voxel = new THREE.Mesh( geo, material );
+            voxel.position.set((block[0]*50)+25, (block[1]*50)+25, (block[2]*50)+25);
+            scene2.add( voxel );
+            objects2.push( voxel );
+            const edges = new THREE.EdgesGeometry( geo );  // outline the blocks for visibility
+            const line = new THREE.LineSegments( edges, new THREE.LineBasicMaterial( { color: 0x000000 } ) );
+            line.position.set((block[0]*50)+25, (block[1]*50)+25, (block[2]*50)+25);
+            scene2.add( line );
+        })
+    }
 
     // lights
     const ambientLight = new THREE.AmbientLight( 0x606060 );
@@ -270,9 +199,7 @@ function onPointerMove( event ) {
             rollOverMesh2.position.copy( intersect.point ).add( intersect.face.normal );
             rollOverMesh2.position.divideScalar( 50 ).floor().multiplyScalar( 50 ).addScalar( 25 );
         }
-        
     }
-
     camera2.position.set( camera1.position.x, camera1.position.y, camera1.position.z );
     camera2.lookAt( 0, 0, 0 );
     render();
