+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/extra_refs/voxel_painter_q3.mjs
@@ -22,10 +22,20 @@ let actions_taken = []; // [original_block, new_block, action_type]
 var startedHIT = false;
 
 let starting_shapes = [
-    [0,0,0,0],
-    [-1,0,0,2],
-    [0,0,-1,4],
-    [-1,0,-1,0]
+    [2,0,2,0],
+    [2,1,2,1],
+    [2,2,2,2],
+    [1,0,2,3],
+    [0,0,2,4],
+    [-1,0,2,5],
+    [-2,0,-2,0],
+    [-2,1,-2,1],
+    [-2,0,-1,2],
+    [-2,1,-1,3],
+    [-1,0,-2,4],
+    [-1,1,-2,5],
+    [-1,0,-1,4],
+    [-1,1,-1,5]
 ];
 
 init1();
