+++ droidlet/tools/crowdsourcing/vision_annotation_task/server_files/extra_refs/voxel_wrapper_q4.html
@@ -15,17 +15,17 @@
         </div>
         <input type="hidden" id="markedBlocks" name="markedBlocks" value="">
     </body>
-    <script type='module' id="voxelPainter" src='voxel_painter_q1.mjs'></script>
+    <script type='module' src='voxel_painter_q4.mjs'></script>
     <script>
         document.getElementById("finish-annotation").addEventListener("click", function() {
-            let output_obj = {q1_output: document.getElementById("markedBlocks").value};
+            let output_obj = {q4_output: document.getElementById("markedBlocks").value};
             window.parent.postMessage(JSON.stringify({ msg: output_obj }), "*");
             document.getElementById("finish-annotation").disabled = true;
             document.getElementById("null-result").disabled = true;
             document.getElementById("response").innerHTML = " - Annotation Complete";
         });
         document.getElementById("null-result").addEventListener("click", function() {
-            let output_obj = {q1_output: "null"};  // parser expects a string
+            let output_obj = {q4_output: "null"};  // parser expects a string
             window.parent.postMessage(JSON.stringify({ msg: output_obj }), "*");
             document.getElementById("finish-annotation").disabled = true;
             document.getElementById("null-result").disabled = true;
